% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_api_access.R
\name{check_api_access}
\alias{check_api_access}
\title{Check access to https://api.finna.fi/api/v1/?openapi}
\usage{
check_api_access()
}
\value{
a logical.
}
\description{
Check if R has access to resources at https://api.finna.fi/api/v1/?openapi
}
\examples{
 \dontrun{
   check_api_access()
 }
}
