% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fennica_subset}
\alias{fennica_subset}
\title{Fennica Subset Data}
\format{
A data frame with X rows and Y variables:
\describe{
\item{id}{Unique identifier}
\item{Title}{Title of the work}
\item{Author}{Author(s) of the work}
\item{Year}{Publication year}
\item{Language}{Language of the work}
\item{Formats}{Available formats}
\item{Subjects}{Subject categories}
\item{Library}{Holding library}
\item{Series}{Series information}
\item{last_indexed}{Date of last indexing}
}
}
\source{
\url{https://example.com/fennica}
}
\usage{
fennica_subset
}
\description{
A subset of the Fennica dataset for demonstration purposes.
}
\keyword{datasets}
