% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finna_as_json.R
\name{finna_as_json}
\alias{finna_as_json}
\alias{finna_as_json.finna_query}
\title{Convert Finna query object to JSON}
\usage{
finna_as_json(x)

\method{finna_as_json}{finna_query}(x)
}
\arguments{
\item{x}{An object to convert (e.g., a Finna API query or response object).}
}
\description{
Convert Finna query object to JSON
}
\keyword{internal}
