% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finna_cite.R
\name{finna_cite}
\alias{finna_cite}
\title{Cite a Finna collection}
\usage{
finna_cite(result, index, style = "citation")
}
\arguments{
\item{result}{The Finna collection result as a tibble.}

\item{index}{The index of the collection to cite (numeric).}

\item{style}{The citation style to use (default: "citation"). See \code{\link[utils]{bibentry}}.}
}
\value{
A bibliographic entry (\code{bibentry}) printed in the specified style.
}
\description{
Automatically generates a citation for a Finna collection result.
}
