% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finna_interactive.R
\name{finna_interactive}
\alias{finna_interactive}
\title{Interactive Finna Search and Data Download}
\usage{
finna_interactive()
}
\value{
A dataframe containing the selected dataset or downloaded data.
}
\description{
Provides an interactive interface to search, select, and download datasets from Finna API.
}
\seealso{
\code{\link[=search_finna]{search_finna()}}, \code{\link[=fetch_finna]{fetch_finna()}}, \code{\link[=finna_cite]{finna_cite()}}
}
