% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_all_finna_data.R
\name{get_all_finna_data}
\alias{get_all_finna_data}
\title{Retrieve All Finna Data with Pagination, Sorting, and Provenance Information}
\usage{
get_all_finna_data(
  lookfor = "",
  fields = NULL,
  sort = "relevance,id asc",
  limit = 100,
  lng = "fi",
  max_pages = 1e+05
)
}
\arguments{
\item{lookfor}{A string containing the search terms. Boolean operators (AND, OR, NOT) can be included.}

\item{fields}{A vector of fields to be returned in the search results. Defaults to NULL, which returns a standard set of fields.}

\item{sort}{A string defining the sort order of the results. Defaults to "relevance,id asc". Options include:
\itemize{
\item "relevance,id asc" (default)
\item "main_date_str desc" (Year, newest first)
\item "main_date_str asc" (Year, oldest first)
\item "last_indexed desc" (Last modified)
\item "first_indexed desc" (Last added)
\item "callnumber,id asc" (Classmark)
\item "author,id asc" (Author)
\item "title,id asc" (Title)
}}

\item{limit}{An integer specifying the number of records to return per page. Defaults to 100 (maximum).}

\item{lng}{A string for the language of returned translated strings. Options are "fi", "en-gb", "sv", "se". Defaults to "fi".}

\item{max_pages}{An integer indicating the maximum number of pages to retrieve. Default is 10, but can be increased.}
}
\value{
A tibble containing all the search results, without provenance or citation information.
}
\description{
This function retrieves all available Finna data incrementally by paginating through the API results,
with the option to sort the results and append data citation and provenance information to the output.
}
\examples{
all_finna_data <- get_all_finna_data(lookfor = "sibelius", sort = "main_date_str desc", limit = 100, max_pages = 5)
print(all_finna_data)
}
