% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_finna.R
\name{get_finna}
\alias{get_finna}
\title{Get Finna Data}
\usage{
get_finna(
  lookfor = "",
  type = "AllFields",
  fields = NULL,
  filters = NULL,
  sort = "relevance,id asc",
  page = 1,
  limit = 100,
  lng = "fi",
  cache = TRUE,
  cache_dir = NULL,
  compress_file = TRUE,
  use.data.table = FALSE,
  prettyPrint = FALSE
)
}
\arguments{
\item{lookfor}{A string containing the search terms. Boolean operators (AND, OR, NOT) can be included.}

\item{type}{A string specifying the type of search. Options include "AllFields", "Title", "Author", "Subject". Defaults to "AllFields".}

\item{fields}{A vector of fields to be returned in the search results. Defaults to NULL, which returns a standard set of fields.}

\item{filters}{A named list of filters. Names of list objects are Finna variable codes, and values are vectors of observation codes.}

\item{sort}{A string defining the sort order of the results. Options include "relevance,id asc" (default), "main_date_str desc", "main_date_str asc", and more.}

\item{page}{An integer indicating the page number of results to retrieve. Defaults to 1.}

\item{limit}{An integer specifying the number of records to return per page. Defaults to 100.}

\item{lng}{A string for the language of returned translated strings. Options are "fi", "en-gb", "sv", "se". Defaults to "fi".}

\item{cache}{A logical indicating whether to cache the results. Defaults to TRUE.}

\item{cache_dir}{A path to a cache directory. NULL (default) uses and creates a 'finna' directory in the temporary directory defined by base R \code{\link[=tempdir]{tempdir()}} function.}

\item{compress_file}{A logical whether to compress the RDS-file in caching. Default is TRUE.}

\item{use.data.table}{Logical. Should data.table functions be used? Default is FALSE.}

\item{prettyPrint}{A logical value indicating whether to pretty-print the JSON response. Useful for debugging. Defaults to FALSE.}
}
\value{
A tibble containing the search results with relevant fields extracted.
}
\description{
Download data sets from Finna \url{https://www.finna.fi}
}
\examples{
finna_data <- get_finna("sibelius", limit = 100, max_pages = 5)
print(finna_data)
}
