% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrateFinna.R
\name{integrate_metadata}
\alias{integrate_metadata}
\title{Integrate Finna Metadata with Another Dataset}
\usage{
integrate_metadata(metadata1, metadata2, key = "Title")
}
\arguments{
\item{metadata1}{A tibble containing refined Finna metadata.}

\item{metadata2}{A tibble containing another dataset to merge with the Finna metadata.}

\item{key}{A string specifying the key to join by. Defaults to "Title".}
}
\value{
A tibble containing the merged dataset.
}
\description{
Merges Finna metadata with another dataset using a common key (e.g., "Title").
}
\examples{
library(dplyr)
finna_data <- search_finna("sibelius")

# Example other dataset to merge with
other_data <- tibble::tibble(
  Title = c("Sibelius Symphony No. 5", "Finlandia", "Valse Triste"),
  Rating = c(5, 4, 3)
)

# Integrate the two datasets by "Title"
integrated_data <- integrate_metadata(finna_data, other_data, key = "Title")
print(integrated_data)
}
