% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_finna_from_file.R
\name{search_finna_from_file}
\alias{search_finna_from_file}
\title{Search Finna using text from a file (extended version)}
\usage{
search_finna_from_file(file_path, limit = 10, lng = "fi", query_limit = 500)
}
\arguments{
\item{file_path}{The path to the text file.}

\item{limit}{The number of results to return from Finna. Defaults to 10.}

\item{lng}{Language for returned translated strings. Defaults to "fi".}

\item{query_limit}{The maximum length of the query string allowed for the API. Defaults to 500 characters.}
}
\value{
A tibble containing the Finna search results.
}
\description{
Reads text from a file, processes it, and searches it in Finna using the \code{search_finna} function.
}
