% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualizeFinna.R
\name{visualize_author_distribution}
\alias{visualize_author_distribution}
\title{Visualize Distribution by Authors}
\usage{
visualize_author_distribution(metadata)
}
\arguments{
\item{metadata}{A tibble containing refined Finna metadata, with an "Author" column.}
}
\value{
A ggplot2 object showing the distribution of records by authors.
}
\description{
Creates a bar plot showing the distribution of records by authors.
}
\examples{
library(finna)
sibelius_data <- search_finna("sibelius")
refined_data <- refine_metadata(sibelius_data)
visualize_author_distribution(refined_data)
}
