% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualizeFinna.R
\name{visualize_format_distribution}
\alias{visualize_format_distribution}
\title{Visualize Distribution by Formats}
\usage{
visualize_format_distribution(metadata)
}
\arguments{
\item{metadata}{A tibble containing refined Finna metadata, with a "Formats" column.}
}
\value{
A ggplot2 object showing the distribution of records by formats.
}
\description{
Creates a bar plot showing the distribution of records by their formats.
}
\examples{
library(finna)
sibelius_data <- search_finna("sibelius")
refined_data <- refine_metadata(sibelius_data)
visualize_format_distribution(refined_data)
}
