% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualizeFinna.R
\name{visualize_library_distribution}
\alias{visualize_library_distribution}
\title{Visualize Distribution by Libraries}
\usage{
visualize_library_distribution(metadata)
}
\arguments{
\item{metadata}{A tibble containing refined Finna metadata, with a "Library" column.}
}
\value{
A ggplot2 object showing the distribution of records by libraries.
}
\description{
Creates a bar plot showing the distribution of records by libraries.
}
\examples{
library(finna)
sibelius_data <- search_finna("sibelius")
refined_data <- refine_metadata(sibelius_data)
visualize_library_distribution(refined_data)
}
