% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize_finna.R
\name{visualize_subject_distribution}
\alias{visualize_subject_distribution}
\title{Visualize Distribution by Subjects}
\usage{
visualize_subject_distribution(metadata)
}
\arguments{
\item{metadata}{A tibble containing refined Finna metadata, with a "Subjects" column.}
}
\value{
A ggplot2 object showing the distribution of records by subjects.
}
\description{
Creates a bar plot showing the distribution of records by subjects.
}
\examples{
library(finna)
sibelius_data <- search_finna("sibelius")
refined_data <- refine_metadata(sibelius_data)
visualize_subject_distribution(refined_data)
}
