% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualizeFinna.R
\name{visualize_year_distribution_line}
\alias{visualize_year_distribution_line}
\title{Visualize Year Distribution (Line Plot)}
\usage{
visualize_year_distribution_line(metadata)
}
\arguments{
\item{metadata}{A tibble containing refined Finna metadata, with a "Year" column.}
}
\value{
A ggplot2 object representing the line plot.
}
\description{
Creates a line plot showing the distribution of records by year.
}
\examples{
library(finna)
sibelius_data <- search_finna("sibelius")
refined_data <- refine_metadata(sibelius_data)
visualize_year_distribution_line(refined_data)
}
