% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fennica_subset}
\alias{fennica_subset}
\title{Fennica Subset Data}
\format{
A data frame with 54749 rows and 6 variables:
\describe{
\item{ID}{Unique identifier}
\item{Author}{Author(s) of the work, including MARC encoding}
\item{Name_100a}{Primary name entry for the author}
\item{Code_1000}{Authority control code for the author}
\item{Code_7000}{Additional linked authors or contributors}
\item{Code_7001}{Authority control code for secondary contributors}
}
}
\source{
\url{https://www.finna.fi/}
}
\usage{
fennica_subset
}
\description{
A subset of the Fennica dataset for demonstration purposes.
}
\keyword{datasets}
