% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finna_interactive.R
\name{finna_interactive}
\alias{finna_interactive}
\title{Interactive function for Finna Search}
\usage{
finna_interactive(query = NULL)
}
\arguments{
\item{query}{A search query string for finding the dataset of interest.}
}
\description{
A simple interactive function, that helps with searching and downloading Finna data.
}
\seealso{
\code{\link[=search_finna]{search_finna()}}
}
