% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refinemetaData.R
\name{refine_metadata}
\alias{refine_metadata}
\title{Refine Finna Metadata}
\usage{
refine_metadata(data)
}
\arguments{
\item{data}{A tibble containing raw Finna metadata.}
}
\value{
A tibble with selected, cleaned metadata fields, or NULL if required fields are missing.
}
\description{
Refines the Finna metadata tibble by keeping relevant fields and cleaning up missing values.
}
\examples{
library(finna)
sibelius_data <- search_finna("sibelius")
refine_metadata(sibelius_data)

}
